requirements.txt
appdirs==1.4.3
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cycler==0.10.0
Cython==0.25.2
decorator==4.1.2
Django==1.11.5
easydict==1.7
enum==0.4.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
html5lib==0.9999999
idna==2.6
image==1.5.15
imageio==2.2.0
Markdown==2.6.9
matplotlib==2.0.2
mock==2.0.0
networkx==2.0
numpy==1.13.1
olefile==0.44
packaging==16.8
pbr==3.1.1
Pillow==4.1.1
protobuf==3.4.0
pycparser==2.18
pyglet==1.3.0
pyparsing==2.2.0
pypng==0.0.18
python-dateutil==2.6.0
pytube==6.4.2
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
requests==2.18.4
scikit-image==0.13.1
scipy==0.18.1
six==1.10.0
ssim==0.1.26
subprocess32==3.2.7
tensorboardX==0.8
torch==0.3.1
torchvision==0.2.0
urllib3==1.22
Werkzeug==0.12.2
